Splits extensions into multiple smaller files
